Maven publication: use the new `maven-publish` plugin
We have to do cumbersome tricks to get dependencies correct (especially in the generated POM), but all in all we have now better control of the generated packages. And, more importantly, we can support several flavors, publishing each flavor to a dedicated artifact ID.

NOTE: So far, the “offline” flavor targets a snapshot version of the SDK (not released yet).