Share `Index` instances
Each `Client` keeps a map of already created indices, with the index name as the key and a weak reference to the index as a value. When asked for an index with the same name, it returns the already created one, unless it has been garbage-collected.

NOTE: In the tests, running the garbage collector does not finalize the index, although it ought to work (it does on a simple example). However, the debugger is positive that the only reference to the object is the weak map. => We’ll assume it should work in the real world… :/