Add an option to choose a different executor for completion handlers
By default, the main thread is used.

This required quite a bit of refactoring, because `AsyncTask` doesn’t allow running the completion on anything else than the UI thread! I documented the reasons in the `Request` interface (private documentation).

I took the opportunity to move the base asynchronous task to a top-level class. There is still a specialized class internal to `AbstractClient` to provide easy access to the client’s properties and methods.

The offline mode also had to be refactored, because now that the completion executor can be changed to a potentially different thread, or even a pool of threads, we have to synchronize the fallback logic.

Fixes #92.